Fix symbolic links creation
The symbolic links are created from the project top level directory
and point to a subdirectory directly below it. But they are created in
the `checkouts` subdirectory, so they point to the wrong place and
they are broken.

We need to create them relative to the parent directory of `checkouts`
sub directory.